+++ src/changelog.md
@@ -2,6 +2,8 @@
 * Requires AzureStack Update 2203+
 ## Azs.Compute.Admin module
 * Get-AzsComputeScaleUnit - Added admin support to view the tenant VM and compute node information of a target scale unit
+## AzS.Network.Admin module
+* Get-AzsVirtualNetworkGatewayConnection - Added support for admin to view all VPN resources created on the stamp
 
 # Release 2.2.0
 * Requires AzureStack Update 2108+